pyproject.toml
[tool.poetry]
name = "raspberry-pi-os"
version = "0.1.0"
description = ""
authors = ["Jono Hill <jono@hillnz.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
jsonpath-ng = "^1.5.2"
GitPython = "^3.1.12"
PyGithub = "^1.54.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
